+++ relay-server/src/services/store.rs
@@ -274,6 +274,12 @@ impl StoreService {
                 ItemType::Span => {
                     self.produce_span(scoping, start_time, event_id, retention, item)?
                 }
+                ItemType::ProfileChunk => self.produce_profile_chunk(
+                    scoping.organization_id,
+                    scoping.project_id,
+                    start_time,
+                    item,
+                )?,
                 other => {
                     let event_type = event_item.as_ref().map(|item| item.ty().as_str());
                     let item_types = envelope
@@ -1035,6 +1041,27 @@ impl StoreService {
             }
         }
     }
+
+    fn produce_profile_chunk(
+        &self,
+        organization_id: u64,
+        project_id: ProjectId,
+        start_time: Instant,
+        item: &Item,
+    ) -> Result<(), StoreError> {
+        let message = ProfileChunkKafkaMessage {
+            organization_id,
+            project_id,
+            received: UnixTimestamp::from_instant(start_time).as_secs(),
+            payload: item.payload(),
+        };
+        self.produce(
+            KafkaTopic::Profiles,
+            organization_id,
+            KafkaMessage::ProfileChunk(message),
+        )?;
+        Ok(())
+    }
 }
 
 impl Service for StoreService {
@@ -1405,6 +1432,14 @@ struct MetricsSummaryKafkaMessage<'a> {
     tags: BTreeMap<&'a str, &'a str>,
 }
 
+#[derive(Clone, Debug, Serialize)]
+struct ProfileChunkKafkaMessage {
+    organization_id: u64,
+    project_id: ProjectId,
+    received: u64,
+    payload: Bytes,
+}
+
 #[derive(Debug, Serialize)]
 #[serde(transparent)]
 struct CogsKafkaMessage(Vec<u8>);
@@ -1436,6 +1471,7 @@ enum KafkaMessage<'a> {
     },
     MetricsSummary(MetricsSummaryKafkaMessage<'a>),
     Cogs(CogsKafkaMessage),
+    ProfileChunk(ProfileChunkKafkaMessage),
 }
 
 impl Message for KafkaMessage<'_> {
@@ -1453,6 +1489,7 @@ impl Message for KafkaMessage<'_> {
             KafkaMessage::Span { .. } => "span",
             KafkaMessage::MetricsSummary(_) => "metrics_summary",
             KafkaMessage::Cogs(_) => "cogs",
+            KafkaMessage::ProfileChunk(_) => "profile_chunk",
         }
     }
 
@@ -1476,7 +1513,8 @@ impl Message for KafkaMessage<'_> {
             | Self::ReplayRecordingNotChunked(_)
             | Self::Span { .. }
             | Self::MetricsSummary(_)
-            | Self::Cogs(_) => Uuid::nil(),
+            | Self::Cogs(_)
+            | Self::ProfileChunk(_) => Uuid::nil(),
 
             // TODO(ja): Determine a partitioning key
             Self::Metric { .. } => Uuid::nil(),